requirements.txt
channels==3.0.3
channels-redis==3.2.0
psycopg2-binary==2.8.6
Django==3.1.12
django-extensions==3.1.3
djangorestframework==3.12.4
drf-extensions==0.7.0
drf-yasg[validation]==1.20.0
django-model-utils==4.1.1
django-phonenumber-field==5.2.0
phonenumbers==8.12.27
factory-boy==3.2.0
celery[redis]==4.4.7
mock==4.0.3
iso8601==0.1.16
django-allauth==0.45.0
django-rest-auth[with_social]==0.9.5
sentry-sdk==1.3.0
redis==3.5.3
django-redis==4.12.1
django-constance[redis]==2.8.0
django-click==2.2.0
Pillow==8.3.1
requests[security]==2.26.0
sorl-thumbnail==12.7.0
django-cors-headers==3.7.0
django-ckeditor==6.1.0
django-push-notifications==2.0.0
django-mptt==0.12.0
python-slugify==5.0.2
user-agents==2.2.0
graypy==2.1.0

django-easy-audit==1.3.0

#num2words==0.5.10

# for deploy
gunicorn==20.0.4
Fabric==2.5.0
asgiref==3.2.10

# dev
ipython
django-admin-csvexport==1.9
